Calcul d'un score global

Permet de calculer un score global à partir d'un bilan qualité

# S3 method for QA_matrix
compute_score(x, score_pond = c(qs_residual_sa_on_sa
  = 6.66, f_residual_sa_on_sa = 6.68, combined_residual_sa_on_sa = 6.66,
  combined_residual_sa_on_sa_last_years = 15, combined_residual_sa_on_i =
  3.33, qs_residual_sa_on_i = 3.33, f_residual_sa_on_i = 3.34,
  f_residual_td_on_sa = 10, f_residual_td_on_i = 10, residuals_mean = 10,
  residuals_independency = 10, residuals_homoskedasticity = 5,
  residual_normality = 5, m7 = 3, q_m2 = 2), modalities = c("Good",
  "Uncertain", "", "Bad", "Severe"), normalize_score_value,
  na.rm = FALSE, n_contrib_score, ...)

Arguments

x
objet de type QA_matrix ou mQA_matrix.

score_pond
formule utilisée pour calculer le score global

modalities
modalités triées par leur ordre d'importance dans le calcul du score (voir détails).

normalize_score_value
chiffre indiquant la valeur de référence pour la normalisation des pondérations utilisées lors du calcul du score. Si le paramètre n'est pas renseigné alors les poids ne seront pas normalisés.

na.rm
booléen indiquant si les valeurs manquantes doivent être enlevées pour le calcul du score.

n_contrib_score
entier indiquant le nombre de variables à créer dans la matrice des valeurs du bilan qualité contenant les n_contrib_score plus grandes contributrices au score (voir détails). si non spécifié alors aucune variable n'est créée.

...
autres paramètres non utilisés.

Value

Un objet de type QA_matrix ou mQA_matrix.

Details

La fonction compute_score permet de calculer un score à partir des modalités d'un bilan qualité. Pour cela chaque modalité est associée à un poids défini par le paramètre modalities. Ainsi, le paramètre par défaut étant c("Good", "Uncertain", "Bad","Severe"), la valeur "Good" sera associée à la note 0, la valeur "Uncertain" sera associée à la note 1, la valeur "Bad" sera associée à la note 2 et la valeur "Bad" sera associée à la note 3.

Le calcul du score se fait grâce au paramètre score_pond qui est un vecteur de numeriques nommé contenant des poids et dont les noms correspondent aux variables de la matrice des modalités à utiliser dans le score. Ainsi, avec le paramètre score_pond = c(qs_residual_sa_on_sa = 10, f_residual_td_on_sa = 5) le score sera calculé à partir des deux variables qs_residual_sa_on_sa et f_residual_td_on_sa. Les notes associées aux modalités de la variable qs_residual_sa_on_sa seront multipliées par 10 et celles associées à la variable f_residual_td_on_sa seront multipliées par 5. Dans le calcul du score, certaines variables peuvent être manquantes, pour ne pas prendre en compte ces valeurs dans le calcul il suffit d'utiliser le paramètre na.rm = TRUE.

Le paramètre normalize_score_value permet de normaliser les scores. Par exemple, si l'on souhaite avoir des notes entre 0 et 20 il suffit d'utiliser le paramètre normalize_score_value = 20.

Le paramètre n_contrib_score permet d'ajouter des nouvelles variables à la matrice des valeurs du bilan qualité dont les valeurs correspondent aux noms des variables qui contribuent le plus au score de la série. n_contrib_score est un entier qui est égal au nombre de variables contributrices que l'on souhaite exporter. Par exemple, pour n_contrib_score = 3, trois colonnes seront crées contenant les trois plus grandes contributrices au score. Les noms des nouvelles variables sont *i*_highest_score, *i* correspondant au rang en terme de contribution au score (1_highest_score contiendra les noms des plus grandes contributrices, 2_highest_score des deuxièmes plus grandes contributrices, etc.). Seules les variables qui ont une contribution non nulle au score sont prises en compte. Ainsi, si une série a un score nul alors toutes les colonnes *i*_highest_score associées à cette série seront vides ; si une série a un score positif uniquement du fait de la variable "m7" alors la valeur correspondante de le variable 1_highest_score sera égale à "m7" et celle des autres variables *i*_highest_score seront vides.

See also

Examples

# NOT RUN {
QA <- extract_QA()
QA <- compute_score(QA,n_contrib_score = 2)
QA
QA$modalities$score
# }