Extract a quality assessment matrix

Functions to extracts a quality report assessment from a CSV file exported by JDemetra+ (or the JDemetra+ cruncher) or from a seasonal adjustment object create by the RJDemetra package.

extract_QA(x, ...)

# S3 method for SA
extract_QA(x, series_name, ...)

# S3 method for default
extract_QA(x, sep = ";", dec = ",", ...)

Arguments

x
a CSV file containing the diagnostics matrix or a SA object. By default a dialog box opens to select a CSV file.

...
unused argument.

series_name
The name of series for the extract_QA.SA method. By default the name of the x parameter is used.

sep
the field separator character in the CSV file (sep = ";" by default).

dec
the character used in the file for decimal points (dec = "," by default).

Value

A QA_matrix object.

Details

The function allows to extract a quality assessment matrix from a CSV file containing all the diagnostics (usually the demetra_m.csv file) or from a SA object create by the package RJDemetra).

The result is a QA_matrix object which is a list of three parameters: * le paramètre modalities est un data.frame contenant un ensemble de variables sous forme catégorielle (Good, Uncertain, Bad, Severe). * le paramètre values est un data.frame contenant les valeurs associées aux indicateurs présents dans modalities (i.e. : p-valeurs, statistiques, etc.) ainsi que des variables qui n'ont pas de modalité (fréquence de la série et modèle ARIMA). * le paramètre score_formula est initié à NULL : il contiendra la formule utilisée pour calculer le score (si le calcul est fait).

See also

Examples

# NOT RUN {
QA <- extract_QA()
QA
# To extract the modalities' matrix:
QA$modalities
# Ou :
QA[["modalities"]]
# }